Move test to separate file
